Create tombsearchresults.php
There are two ways to search for tombs on noladeadspace.com. First, you can choose a tomb from an interactive map. Second, A table is output that has the map tomb number, family name of the tomb, arch diocese number, and grid location on the map. Family name is hypertext leading to a search results page that outputs a complex table containing all data collected on the project. While the table containing all tombs was constructed using the Views module, the more detailed table for each tomb is a bit complex and required custom php code. this is the php code.